+++ lib/blueprint.dart
@@ -50,38 +50,42 @@ class _BlueprintState extends State<Blueprint> with WidgetsBindingObserver {
         init: controller,
         autoRemove: false,
         builder: (controller) {
-          return Container(
-            color: Colors.grey.shade900,
-            key: controller.widgetKey,
-            child: InteractiveViewer(
-              transformationController: controller.transformationController,
-              clipBehavior: Clip.antiAlias,
-              onInteractionStart: controller.onInteractionStart,
-              minScale: 0.1,
-              maxScale: 1.0,
-              // boundaryMargin: const EdgeInsets.all(32),
-              constrained: false,
-              child: Container(
-                color: Colors.grey.shade900,
-                width: controller.size.width,
-                height: controller.size.height,
-                child: Stack(
-                  key: controller.stackKey,
-                  fit: StackFit.expand,
-                  children: [
-                    Positioned.fill(
-                      child: AnimatedOpacity(
-                        opacity: controller.showGrid ? 1.0 : 0.0,
-                        duration: 200.milliseconds,
-                        child: GridPaper(
-                          color: Colors.white.withOpacity(0.1),
-                          divisions: 2,
-                          interval: 100,
+          return Listener(
+            onPointerMove: (event) {
+              controller.pointerPosition = event.localPosition;
+            },
+            child: Container(
+              color: Colors.grey.shade900,
+              key: controller.widgetKey,
+              child: InteractiveViewer(
+                transformationController: controller.transformationController,
+                clipBehavior: Clip.antiAlias,
+                onInteractionStart: controller.onInteractionStart,
+                minScale: 0.1,
+                maxScale: 1.0,
+                constrained: false,
+                child: Container(
+                  color: Colors.grey.shade900,
+                  width: controller.size.width,
+                  height: controller.size.height,
+                  child: Stack(
+                    key: controller.stackKey,
+                    fit: StackFit.expand,
+                    children: [
+                      Positioned.fill(
+                        child: AnimatedOpacity(
+                          opacity: controller.showGrid ? 1.0 : 0.0,
+                          duration: 200.milliseconds,
+                          child: GridPaper(
+                            color: Colors.white.withOpacity(0.1),
+                            divisions: 2,
+                            interval: 100,
+                          ),
                         ),
                       ),
-                    ),
-                    ...controller.nodes
-                  ],
+                      ...controller.nodes
+                    ],
+                  ),
                 ),
               ),
             ),