+++ lib/widgets/fixed_node/fixed_node_controller.dart
@@ -1,0 +1,5 @@
+import 'package:blueprint_system/widgets/node/node_controller.dart';
+
+class FixedNodeController extends NodeController {
+  FixedNodeController(super.initPosition, super.initSize);
+}